+++ sciencebeam_gym/tools/vocabulary/extract_embedding_vocabulary.py
